requirements.txt
httpx
yarl
pycryptodomex
websocket-client
git+https://www.github.com/justfoolingaround/fzf.py